requirements.txt
# Job Pipeline Tracker Bot - Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
schedule>=1.2.0
selenium>=4.15.0
webdriver-manager>=4.0.0
resend>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
